refactor: move external utils to util/ dir
